# mbia-specs/technical/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mbia API
  version: 0.2.0
  summary: REST contract for the Mbia commercial MVP
  description: |
    Contract-first API for Mbia, a private collaborative family heritage platform.

    Product specifications are authoritative for business behavior. This OpenAPI
    document is the authoritative HTTP contract derived from those specifications.
servers:
  - url: https://api.mbia.example.com/api/v1
    description: Production placeholder
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /families/{familyId}/persons:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
    get:
      tags: [Persons]
      operationId: searchPersons
      summary: Search or list persons in a family
      description: |
        Returns ACTIVE Persons by default, ordered by display name (locale-independent,
        case- and accent-folded), then creation date, then id. `status=ARCHIVED` lists archived
        Persons instead, so that an ADMIN can restore them; it is ADMIN only.
        MERGED Persons are never listed.
      parameters:
        - name: status
          in: query
          required: false
          description: ACTIVE (default) or ARCHIVED. ARCHIVED requires the ADMIN role (403 otherwise).
          schema:
            type: string
            enum: [ACTIVE, ARCHIVED]
            default: ACTIVE
        - name: search
          in: query
          required: false
          description: Case- and accent-insensitive substring search, after trimming, in first name, last name, preferred name and "first name last name".
          schema:
            type: string
            maxLength: 200
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Persons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonPage'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Persons]
      operationId: createPerson
      summary: Create a person in the family
      description: |
        ADMIN or CONTRIBUTOR only. If the server detects likely duplicates and
        `confirmPossibleDuplicate` is false, the server returns 409 with code
        `POSSIBLE_DUPLICATE` and candidate summaries in `details.candidates`
        (at most 10 objects shaped as `PersonSummary`, in search order).
        With `linkToCurrentUser`, the server returns 409 `USER_ALREADY_LINKED` when the
        current user already has a linked Person in this family; nothing is created.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonRequest'
      responses:
        '201':
          description: Person created
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/persons/{personId}:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Persons]
      operationId: getPerson
      summary: Get a person profile
      description: |
        Returns the Person whatever its status. An ARCHIVED Person stays readable so
        that an ADMIN can open and restore it; a MERGED Person carries
        `mergedIntoPersonId`. Normal lists and the tree only contain ACTIVE Persons.
      responses:
        '200':
          description: Person
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Persons]
      operationId: updatePerson
      summary: Update person identity/profile data
      description: |
        ADMIN or authorized CONTRIBUTOR according to linked-person protection rules.
        A linked person's protected identity can be changed only by that linked user
        or an ADMIN: every field of the request is protected, and another CONTRIBUTOR
        gets 403 `PERMISSION_DENIED`.
        Only an ACTIVE Person can be edited: an ARCHIVED or MERGED Person returns 404
        `PERSON_NOT_FOUND`. A request that changes nothing returns 200 with the
        unchanged Person and version. The photo follows the same rules as the other fields
        (OQ-040).
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePersonRequest'
      responses:
        '200':
          description: Updated person
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/persons/{personId}/archive:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - $ref: '#/components/parameters/PersonId'
    post:
      tags: [Persons]
      operationId: archivePerson
      summary: Archive a person
      description: |
        ADMIN only. A person linked to a user must be unlinked first: 409
        `PERSON_ALREADY_CLAIMED` (OQ-023). Archiving an ARCHIVED person with its current
        version changes nothing and returns 200 with the unchanged person and version
        (OQ-024). A MERGED person returns 404 `PERSON_NOT_FOUND` (OQ-025).
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Archived person
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/persons/{personId}/restore:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - $ref: '#/components/parameters/PersonId'
    post:
      tags: [Persons]
      operationId: restorePerson
      summary: Restore an archived person
      description: |
        ADMIN only. Restoring an ACTIVE person with its current version changes nothing and
        returns 200 with the unchanged person and version (OQ-024). A MERGED person returns
        404 `PERSON_NOT_FOUND` (OQ-025).
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Restored person
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/persons/{personId}/claim:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - $ref: '#/components/parameters/PersonId'
    post:
      tags: [Persons]
      operationId: claimPerson
      summary: Link the current user to this Person as “me”
      description: |
        Any ACTIVE member, VIEWER included. Only an ACTIVE Person can be claimed: an ARCHIVED
        or MERGED Person returns 404 `PERSON_NOT_FOUND`. A Person linked to another user
        returns 409 `PERSON_ALREADY_CLAIMED`; a caller already linked to another Person of the
        family returns 409 `USER_ALREADY_LINKED`. Claiming the Person already linked to the
        caller returns 200 with the unchanged Person and version.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Person claimed
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Persons]
      operationId: unclaimPerson
      summary: Remove the User ↔ Person link
      description: |
        The linked user may unlink themselves, whatever their role. ADMIN may correct an
        erroneous link; any other caller gets 403 `PERMISSION_DENIED`. On a Person linked to
        nobody, an ADMIN gets 200 with the unchanged Person and version. Only an ACTIVE Person
        can be unlinked: an ARCHIVED or MERGED Person returns 404 `PERSON_NOT_FOUND`.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Person unlinked
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/persons/{personId}/merge:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - $ref: '#/components/parameters/PersonId'
    post:
      tags: [Persons]
      operationId: mergePerson
      summary: Merge this duplicate person into a target person
      description: |
        ADMIN only. `personId` is the duplicate/source that will become MERGED.
        `targetPersonId` is retained. Scalar target values win when both sides are
        non-empty; source values fill empty target fields (an absent text, `gender`
        `UNKNOWN`, a date of precision `UNKNOWN`); the target is deceased when either
        Person is. Relationships and memories are transferred and deduplicated
        atomically; the source's linked user moves to a target linked to nobody. The
        target keeps its photo, or takes the source's when it has none; the source keeps
        no photo, and its photo becomes ARCHIVED when the target already had one (OQ-047).
        Refusals: 409 `PERSON_MERGE_CONFLICT` with `details.reason`
        `DIFFERENT_LINKED_USERS`, `SELF_RELATIONSHIP` (the two Persons are linked by an
        ACTIVE relationship) or `PARENTAL_CYCLE`; 404 `PERSON_NOT_FOUND` for a MERGED
        Person or one of another family; 409 `PERSON_NOT_ACTIVE` for an ARCHIVED one;
        400 `VALIDATION_FAILED` when `targetPersonId` is `personId`; 409
        `CONCURRENT_MODIFICATION` when a version is stale.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergePersonRequest'
      responses:
        '200':
          description: Merge completed; response is the retained target person
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/persons/{personId}/history:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Persons]
      operationId: getPersonHistory
      summary: Get presentation-safe change history for a person
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Person history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonHistoryPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token validated by Spring Security.
  headers:
    ETag:
      description: Resource version encoded as an HTTP ETag, for example `"3"`.
      schema:
        type: string
        example: '"3"'
  parameters:
    FamilyId:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PersonId:
      name: personId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Expected resource ETag/version. Stale values return 409 CONCURRENT_MODIFICATION.
      schema:
        type: string
        example: '"3"'
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Forbidden:
      description: Authenticated user is not authorized for this family/action
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found in the accessible scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Domain or optimistic-concurrency conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    PersonSummary:
      type: object
      required: [id, familyId, firstName, gender, birth, isDeceased, death, status, version]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        firstName:
          type: string
        middleNames:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        preferredName:
          type: [string, 'null']
        displayName:
          type: string
        gender:
          $ref: '#/components/schemas/Gender'
        birth:
          $ref: '#/components/schemas/PartialDate'
        isDeceased:
          type: boolean
        death:
          $ref: '#/components/schemas/PartialDate'
        profilePictureUrl:
          type: [string, 'null']
          format: uri
          description: Short-lived or CDN URL; never the internal storage key.
        linkedUserId:
          type: [string, 'null']
          format: uuid
        status:
          $ref: '#/components/schemas/PersonStatus'
        relationshipToCurrentUser:
          description: |
            What this Person is to the current user's linked Person
            (kinship from = linked Person, to = this Person). Null when the
            current user has no linked Person in this family.
          oneOf:
            - $ref: '#/components/schemas/KinshipCode'
            - type: 'null'
        version:
          type: integer
          format: int64
          minimum: 0
    PersonResponse:
      allOf:
        - $ref: '#/components/schemas/PersonSummary'
        - type: object
          required: [biography, createdAt, updatedAt]
          properties:
            biography:
              type: [string, 'null']
            mergedIntoPersonId:
              type: [string, 'null']
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CreatePersonRequest:
      type: object
      required: [firstName]
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 150
        middleNames:
          type: [string, 'null']
          maxLength: 250
        lastName:
          type: [string, 'null']
          maxLength: 150
        preferredName:
          type: [string, 'null']
          maxLength: 150
        gender:
          $ref: '#/components/schemas/Gender'
        birth:
          $ref: '#/components/schemas/PartialDate'
        isDeceased:
          type: boolean
          default: false
        death:
          $ref: '#/components/schemas/PartialDate'
        biography:
          type: [string, 'null']
          maxLength: 10000
        profileMediaAssetId:
          type: [string, 'null']
          format: uuid
          description: |
            A READY `PROFILE_PICTURE` asset, uploaded by the caller and not yet used, that
            becomes the Person's photo (OQ-040). Otherwise 409 `MEDIA_NOT_READY` or
            `MEDIA_ALREADY_USED`, 403 for another member's asset, or 404 `MEDIA_NOT_FOUND`
            (OQ-036).
        linkToCurrentUser:
          type: boolean
          default: false
          description: Used by onboarding “Commencer par moi”. Same claim rules apply.
        confirmPossibleDuplicate:
          type: boolean
          default: false
    UpdatePersonRequest:
      type: object
      minProperties: 1
      description: |
        Partial update. An absent or `null` field is left unchanged. A blank string clears
        `middleNames`, `lastName`, `preferredName` or `biography`; `{precision: UNKNOWN}`
        clears a date and `gender: UNKNOWN` the gender. `firstName` cannot be cleared.
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 150
        middleNames:
          type: [string, 'null']
          maxLength: 250
        lastName:
          type: [string, 'null']
          maxLength: 150
        preferredName:
          type: [string, 'null']
          maxLength: 150
        gender:
          $ref: '#/components/schemas/Gender'
        birth:
          $ref: '#/components/schemas/PartialDate'
        isDeceased:
          type: boolean
        death:
          $ref: '#/components/schemas/PartialDate'
        biography:
          type: [string, 'null']
          maxLength: 10000
        profileMediaAssetId:
          type: [string, 'null']
          format: uuid
          description: |
            A READY `PROFILE_PICTURE` asset, uploaded by the caller and not yet used, that
            replaces the current photo (OQ-040). Otherwise 409 `MEDIA_NOT_READY` or
            `MEDIA_ALREADY_USED`, 403 for another member's asset, or 404 `MEDIA_NOT_FOUND`
            (OQ-036).
        removeProfilePicture:
          type: boolean
          description: |
            `true` removes the Person's photo (OQ-040). Sent with `profileMediaAssetId`,
            the request is rejected with 400 `VALIDATION_FAILED`.
    MergePersonRequest:
      type: object
      required: [targetPersonId, sourceVersion, targetVersion]
      properties:
        targetPersonId:
          type: string
          format: uuid
        sourceVersion:
          type: integer
          format: int64
          minimum: 0
        targetVersion:
          type: integer
          format: int64
          minimum: 0
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    PersonPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PersonSummary'
        page:
          $ref: '#/components/schemas/PageMeta'
...
    PersonHistoryPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PersonHistoryEntry'
        page:
          $ref: '#/components/schemas/PageMeta'
...